Geometry: holy shit, finally. Just four months of tinkering. Possibly my longest head-banging session. Zoom and scale at view-filling world scale without having to flip, offset or scale the data itself.
Scroll and zoom solution adapted from WWDC 2012 session on UIScrollView. Scroll and zoom an invisible dummy view, and use the dummy view’s offset+scale to transform the OpenGL scene.

The full transform comes out to:
Offset map data so scaling occurs around the top-left corner (Alaska) to match UIView scaling around (0,0)
Scale the data by zoom, while also flipping the Y axis.
Move the scaled data by the UIScrollView offset to frame it correctly.
Apply the above as matrix ops in reverse order.

In addition to that, flip the viewport. This is not the same as flipping the Y axis of the data, since map data is centered on the equator.